chore: cleanup redundant files and ignore deepeval cache
