+++ .deepeval/.deepeval-cache.json
@@ -1,1 +1,1 @@
-{"test_cases_lookup_map": {"{\"actual_output\": \"The Supreme Court ruled in favor of Phelps, upholding the Fourth Circuit's decision.\", \"context\": [\" The case was named after plaintiff Jim Obergefell, who sued the state of Ohio after officials refused to recognize his marriage on the death certificate of his husband.\", \"On April 28, 2015, the Supreme Court of the United States heard oral arguments for \\\"Obergefell v. Hodges\\\" (Ohio), which was consolidated with three other same-sex marriage cases from the other states in the Sixth Circuit: \\\"Tanco v. Haslam\\\" (Tennessee), \\\"DeBoer v. Snyder\\\" (Michigan), \\\"Bourke v. Beshear\\\" (Kentucky).\"], \"expected_output\": \"Jim Obergefell\", \"hyperparameters\": null, \"input\": \"Who was the widow affected by the Same-sex marriage decision ruled on by the Sixth Circuit from Ohio?\", \"retrieval_context\": []}": {"cached_metrics_data": [{"metric_data": {"name": "Exact Match", "threshold": 1.0, "success": false, "score": 0.0, "reason": "The actual and expected outputs are different.", "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 1.0, "strict_mode": false, "include_reason": false}}, {"metric_data": {"name": "Tool Correctness", "threshold": 0.5, "success": true, "score": 0.5, "reason": "[\n\t Tool Calling Reason: Incomplete tool usage: missing tools [ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in Ohio\"\n    }\n), ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in the Sixth Circuit\"\n    }\n)]; expected ['Wikipedia', 'Wikipedia'], called ['Wikipedia', 'Python']. See more details above.\n\t Tool Selection Reason: No available tools were provided to assess tool selection criteria\n]\n", "strictMode": false, "evaluationCost": 0, "verboseLogs": "Expected Tools:\n[\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in the Sixth Circuit\"\n        }\n    ),\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in Ohio\"\n        }\n    )\n] \n \nTools Called:\n[\n    ToolCall(\n        name=\"Wikipedia\"\n    ),\n    ToolCall(\n        name=\"Python\"\n    )\n] \n \nAvailable Tools: [] \n \nTool Selection Score: 1.0 \n \nTool Selection Reason: No available tools were provided to assess tool selection criteria"}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false, "evaluation_params": []}}, {"metric_data": {"name": "ToolUsageMetric", "threshold": 0.5, "success": false, "score": 0.375, "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false}}, {"metric_data": {"name": "FactsSimilarityMetric", "threshold": 0.5, "success": true, "score": 0.6, "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false}}]}}}
+{"test_cases_lookup_map": {"{\"actual_output\": \"The Supreme Court ruled in favor of Phelps, upholding the Fourth Circuit's decision.\", \"context\": [\" The case was named after plaintiff Jim Obergefell, who sued the state of Ohio after officials refused to recognize his marriage on the death certificate of his husband.\", \"On April 28, 2015, the Supreme Court of the United States heard oral arguments for \\\"Obergefell v. Hodges\\\" (Ohio), which was consolidated with three other same-sex marriage cases from the other states in the Sixth Circuit: \\\"Tanco v. Haslam\\\" (Tennessee), \\\"DeBoer v. Snyder\\\" (Michigan), \\\"Bourke v. Beshear\\\" (Kentucky).\"], \"expected_output\": \"Jim Obergefell\", \"hyperparameters\": null, \"input\": \"Who was the widow affected by the Same-sex marriage decision ruled on by the Sixth Circuit from Ohio?\", \"retrieval_context\": []}": {"cached_metrics_data": [{"metric_data": {"name": "Exact Match", "threshold": 1.0, "success": false, "score": 0.0, "reason": "The actual and expected outputs are different.", "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 1.0, "strict_mode": false, "include_reason": false}}, {"metric_data": {"name": "AnswerLLMJudgeMetric", "threshold": 0.7, "success": false, "score": 0.0, "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 0.7, "strict_mode": false, "include_reason": false}}, {"metric_data": {"name": "Tool Correctness", "threshold": 0.5, "success": true, "score": 0.5, "reason": "[\n\t Tool Calling Reason: Incomplete tool usage: missing tools [ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in Ohio\"\n    }\n), ToolCall(\n    name=\"Wikipedia\",\n    input_parameters={\n        \"query\": \"Same-sex marriage in the Sixth Circuit\"\n    }\n)]; expected ['Wikipedia', 'Wikipedia'], called ['Wikipedia', 'Python']. See more details above.\n\t Tool Selection Reason: No available tools were provided to assess tool selection criteria\n]\n", "strictMode": false, "evaluationCost": 0, "verboseLogs": "Expected Tools:\n[\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in the Sixth Circuit\"\n        }\n    ),\n    ToolCall(\n        name=\"Wikipedia\",\n        input_parameters={\n            \"query\": \"Same-sex marriage in Ohio\"\n        }\n    )\n] \n \nTools Called:\n[\n    ToolCall(\n        name=\"Wikipedia\"\n    ),\n    ToolCall(\n        name=\"Python\"\n    )\n] \n \nAvailable Tools: [] \n \nTool Selection Score: 1.0 \n \nTool Selection Reason: No available tools were provided to assess tool selection criteria"}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false, "evaluation_params": []}}, {"metric_data": {"name": "ToolUsageMetric", "threshold": 0.5, "success": false, "score": 0.375, "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false}}, {"metric_data": {"name": "FactsSimilarityMetric", "threshold": 0.5, "success": true, "score": 0.6, "strictMode": false, "evaluationCost": 0}, "metric_configuration": {"threshold": 0.5, "strict_mode": false, "include_reason": false}}]}}}
