+++ python/.env.example
@@ -120,7 +120,3 @@ EVAL_DISPLAY_MODE=all
 EVAL_CHAT_MODEL_NAME="ollama:granite4:micro"
 
 
-# langfuse
-LANGFUSE_SECRET_KEY = "sk-lf-380d1695-7c67-4842-aeb1-50dad5f2d989"
-LANGFUSE_PUBLIC_KEY = "pk-lf-8eeb86d6-c102-401d-a2a1-bd81345088c2"
-LANGFUSE_BASE_URL = "http://localhost:3000"
