+++ python/eval/agents/requirement/deep_eval_agent_example.py
@@ -50,8 +50,15 @@ from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.errors import FrameworkError
 
 # --- Local Project Imports ---
-from eval.deep_eval import DeepEvalLLM
-from _utils import EvaluationRow, EvaluationTable, print_evaluation_table
+from eval.deep_eval import (
+    DeepEvalLLM,
+    create_evaluation_table,
+)
+from eval._utils import (
+    EvaluationRow,
+    EvaluationTable,
+    print_evaluation_table,
+)
 
 
 # --- Environment Setup ---
@@ -546,53 +553,6 @@ async def create_rag_test_cases(num_rows: int = 50):
 
     return test_cases
 
-
-
-def create_evaluation_table(eval_results, metrics: List[BaseMetric]) -> EvaluationTable:
-    """
-    Converts DeepEval results into a structured EvaluationTable.
-    """
-    def _metric_name(metric_obj):
-        return getattr(metric_obj, "__name__", None) or metric_obj.__class__.__name__
-
-    metric_names = [_metric_name(m) for m in metrics]
-
-    per_test_results = (
-        getattr(eval_results, "results", None)
-        or getattr(eval_results, "test_results", None)
-        or []
-    )
-
-    if isinstance(eval_results, list):
-        per_test_results = eval_results
-
-    rows = []
-    for idx, test_res in enumerate(per_test_results):
-        metrics_data = (
-            getattr(test_res, "metrics_data", None)
-            or getattr(test_res, "metrics_results", None)
-            or []
-        )
-        
-        # Build map for this specific row
-        metric_success_map = {}
-        for md in metrics_data:
-            md_name = (
-                getattr(md, "metric_name", None)
-                or getattr(md, "name", None)
-                or getattr(md, "__name__", None)
-                or md.__class__.__name__
-            )
-            metric_success_map[md_name] = getattr(md, "success", False)
-
-        # Create structured row
-        row = EvaluationRow(
-            test_case_label=f"Test case {idx + 1}",
-            results={name: metric_success_map.get(name, False) for name in metric_names}
-        )
-        rows.append(row)
-        
-    return EvaluationTable(metric_names=metric_names, rows=rows)
 
 
 @pytest.mark.asyncio
