+++ python/eval/deep_eval.py
@@ -2,12 +2,15 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
-from typing import Any, TypeVar
+from typing import Any, List, TypeVar
 
 from deepeval.key_handler import KEY_FILE_HANDLER, ModelKeyValues
+from deepeval.metrics import BaseMetric
 from deepeval.models import DeepEvalBaseLLM
 from dotenv import load_dotenv
 from pydantic import BaseModel
+
+from ._utils import EvaluationRow, EvaluationTable
 
 from beeai_framework.backend import ChatModel, ChatModelParameters
 from beeai_framework.backend.constants import ProviderName
@@ -61,3 +64,53 @@ class DeepEvalLLM(DeepEvalBaseLLM):
         # pyrefly: ignore [bad-argument-type]
         model = ChatModel.from_name(name, options, **kwargs)
         return DeepEvalLLM(model)
+
+
+def create_evaluation_table(eval_results, metrics: List[BaseMetric]) -> EvaluationTable:
+    """
+    Converts DeepEval results into a structured EvaluationTable.
+    """
+    def _metric_name(metric_obj):
+        return getattr(metric_obj, "__name__", None) or metric_obj.__class__.__name__
+
+    metric_names = [_metric_name(m) for m in metrics]
+
+    per_test_results = (
+        getattr(eval_results, "results", None)
+        or getattr(eval_results, "test_results", None)
+        or []
+    )
+
+    if isinstance(eval_results, list):
+        per_test_results = eval_results
+
+    rows = []
+    for idx, test_res in enumerate(per_test_results):
+        metrics_data = (
+            getattr(test_res, "metrics_data", None)
+            or getattr(test_res, "metrics_results", None)
+            or []
+        )
+        
+        # Build map for this specific row
+        metric_success_map = {}
+        for md in metrics_data:
+            md_name = (
+                getattr(md, "metric_name", None)
+                or getattr(md, "name", None)
+                or getattr(md, "__name__", None)
+                or md.__class__.__name__
+            )
+            metric_success_map[md_name] = getattr(md, "success", False)
+
+        # Create structured row
+        row = EvaluationRow(
+            test_case_label=f"Test case {idx + 1}",
+            results={name: metric_success_map.get(name, False) for name in metric_names}
+        )
+        rows.append(row)
+        
+    return EvaluationTable(metric_names=metric_names, rows=rows)
+
+    
+
